work on emergency stop w/o inline keyboard, some code changes
