+++ bot/conf/__init__.py
@@ -1,0 +1,4 @@
+from .greenhouse_config import *
+from .ext_greenhouse_lib import *
+from .greenhouse_lib_english import *
+from .greenhouse_lib_german import *
