+++ bot/peripherals/__init__.py
@@ -1,0 +1,7 @@
+from .oled import *
+from .temperature import *
+from .dht import *
+from .four_digit import *
+from .monitor import *
+from .stop_and_restart import *
+
