+++ bot/peripherals/four_digit/__init__.py
@@ -1,0 +1,1 @@
+from .display import *
